+++ src/web/006_response.js
@@ -98,23 +98,60 @@ Response.prototype.deserializeHeaders = function() {
 //var isUrlAbsoluteRE = /(:\/\/)|(^[-A-z0-9]*\.[-A-z0-9]*)/; // has :// or starts with ___.___
 Response.prototype.processHeaders = function(request) {
 	var self = this;
+
+	// Construct the base URLd out of the via headers
+	var via = self.parsedHeaders.via;
+	var host_proxy = local.viaToUri(via);
+
+	// Helper function to percent-decode the number of times necesssary
+	var decodeSubURI = function(str) {
+		//
+		for (var i=0; i < via.length; i++)
+			str = decodeURIComponent(str);
+		return str;
+	};
+
+	// Update the link headers
 	if (self.parsedHeaders.link) {
 		self.parsedHeaders.link.forEach(function(link) {
-			// if (isUrlAbsoluteRE.test(link.href) === false)
+			// Convert relative paths to absolute uris
 			if (!local.isAbsUri(link.href))
 				link.href = local.joinRelPath(request.urld, link.href);
-			link.host_domain = local.parseUri(link.href).authority;
+
+			// Extract host domain
+			var host_domain = null;
+			if (host_proxy && link.href.indexOf(host_proxy) === 0) {
+				// A suburi of the proxy? See if its an abs uri
+				var suburi = decodeSubURI(link.href.slice(host_proxy.length)).slice(1);
+				if (/^http(s|l)?:\/\//.test(suburi)) {
+					host_domain = local.parseUri(suburi).authority;
+				}
+			}
+			if (!host_domain) {
+				// Not a proxied request, handle as is
+				host_domain = local.parseUri(link.href).authority;
+			}
+
+			// Set host data
+			Object.defineProperty(link, 'host_domain', { enumerable: false, configurable: true, writable: true, value: host_domain });
 			var peerd = local.parsePeerDomain(link.host_domain);
 			if (peerd) {
-				link.host_user   = peerd.user;
-				link.host_relay  = peerd.relay;
-				link.host_app    = peerd.app;
-				link.host_sid    = peerd.sid;
+				Object.defineProperty(link, 'host_user', { enumerable: false, configurable: true, writable: true, value: peerd.user });
+				Object.defineProperty(link, 'host_relay', { enumerable: false, configurable: true, writable: true, value: peerd.relay });
+				Object.defineProperty(link, 'host_app', { enumerable: false, configurable: true, writable: true, value: peerd.app });
+				Object.defineProperty(link, 'host_sid', { enumerable: false, configurable: true, writable: true, value: peerd.sid });
 			} else {
 				delete link.host_user;
 				delete link.host_relay;
 				delete link.host_app;
 				delete link.host_sid;
+			}
+
+			// Add proxy
+			if (host_proxy) {
+				Object.defineProperty(link, 'host_proxy', { enumerable: false, configurable: true, writable: true, value: host_proxy });
+			} else {
+				delete link.host_proxy;
 			}
 		});
 	}